# Refine stack op to improve xlnet performance, test=develop