--- a/paddle/fluid/operators/stack_op.cu
+++ b/paddle/fluid/operators/stack_op.cu
@@ -13,36 +13,16 @@
 // limitations under the License.
 
 #include <algorithm>
+#include <limits>
 #include <vector>
 #include "paddle/fluid/operators/stack_op.h"
+#include "paddle/fluid/platform/gpu_launch_param_config.h"
 
 namespace plat = paddle::platform;
 namespace ops = paddle::operators;
 
 namespace paddle {
 namespace operators {
-
-static inline void Get2DGpuLaunchConfig(
-    const platform::CUDADeviceContext& context, int num_rows, int num_cols,
-    dim3* block_dims, dim3* grid_dims) {
-  // Set the thread block and grid according to CurrentDeviceId
-  const int kThreadsPerBlock = 512;
-  int block_cols = kThreadsPerBlock;
-  if (num_cols < kThreadsPerBlock) {  // block_cols is aligned by 32.
-    block_cols = ((num_cols + 31) >> 5) << 5;
-  }
-  int block_rows = kThreadsPerBlock / block_cols;
-  *block_dims = dim3(block_cols, block_rows, 1);
-
-  int max_threads = context.GetMaxPhysicalThreadCount();
-  int max_blocks = std::max(max_threads / kThreadsPerBlock, 1);
-
-  int grid_cols =
-      std::min((num_cols + block_cols - 1) / block_cols, max_blocks);
-  int grid_rows =
-      std::min(max_blocks / grid_cols, std::max(num_rows / block_rows, 1));
-  *grid_dims = dim3(grid_cols, grid_rows, 1);
-}
 
 template <typename T, typename IntType>
 __global__ void stack_kernel(T** input_ptrs, int split_size, int rows, int cols,
@@ -97,46 +77,30 @@ class StackGPUKernel : public framework::OpKernel<T> {
     x_col = x[0]->numel() / x_row;
     int out_col = x_col * n;
 
-    dim3 block_dims;
-    dim3 grid_dims;
-    Get2DGpuLaunchConfig(dev_ctx, x_row, out_col, &block_dims, &grid_dims);
-    stack_kernel<T, int><<<grid_dims, block_dims, 0, dev_ctx.stream()>>>(
-        reinterpret_cast<T**>(tmp_x_data->ptr()), x_col, x_row, out_col,
-        y_data);
+    auto config = Get2DGpuLaunchConfig(dev_ctx, out_col, x_row);
+
+    if (y->numel() < std::numeric_limits<int32_t>::max()) {
+      stack_kernel<T, int32_t><<<config.block_per_grid, config.thread_per_block,
+                                 0, dev_ctx.stream()>>>(
+          reinterpret_cast<T**>(tmp_x_data->ptr()), x_col, x_row, out_col,
+          y_data);
+    } else {
+      stack_kernel<T, int64_t><<<config.block_per_grid, config.thread_per_block,
+                                 0, dev_ctx.stream()>>>(
+          reinterpret_cast<T**>(tmp_x_data->ptr()), x_col, x_row, out_col,
+          y_data);
+    }
   }
 };
-
-inline int DivUp(int a, int b) { return (a + b - 1) / b; }
-
-static inline void Get1DGpuLaunchConfig(
-    const platform::CUDADeviceContext& context, int element_count,
-    int* threads_per_block, int* blocks_per_grid) {
-  const int theory_thread_count = element_count;
-  int max_threads = context.GetMaxPhysicalThreadCount();
-  const int physical_thread_count = std::min(max_threads, theory_thread_count);
-
-  // Need get from device
-  const int thread_per_block = 1024;
-  *threads_per_block = thread_per_block;
-  // suppose SM can hold 8 block
-  int sm_block = context.GetSMCount() * 8;
-  const int block_count =
-      std::min(DivUp(physical_thread_count, thread_per_block), sm_block);
-  *blocks_per_grid = block_count;
-}
-
-#define CUDA_1D_KERNEL_LOOP(i, n)                              \
-  for (int i = blockIdx.x * blockDim.x + threadIdx.x; i < (n); \
-       i += blockDim.x * gridDim.x)
 
 template <typename T>
 __global__ void unstack_kernel(const T* __restrict__ input, int pre_dim_size,
                                int split_dim_size, int suf_dim_size,
                                int num_split, T** output_ptrs) {
-  eigen_assert(blockDim.y == 1);
-  eigen_assert(blockDim.z == 1);
+  assert(blockDim.y == 1);
+  assert(blockDim.z == 1);
   // In this case they are equal
-  eigen_assert(split_dim_size % num_split == 0);
+  assert(split_dim_size % num_split == 0);
 
   int size = pre_dim_size * split_dim_size * suf_dim_size;
   int each_dim_size = split_dim_size / num_split;
@@ -165,17 +129,22 @@ class StackGradGPUKernel : public framework::OpKernel<T> {
     if (axis < 0) axis += dy->dims().size();
 
     int n = dy->dims()[axis];
-    PADDLE_ENFORCE_EQ(n, dx.size(), "Output dx size should be equal to n");
+    PADDLE_ENFORCE_EQ(n, dx.size(),
+                      platform::errors::InvalidArgument(
+                          "Output dx size should be equal to n, but"
+                          " received n is:%d dx size is:%d.",
+                          n, dx.size()));
+
     // dx is output, so save each data address, then copy each dy into dx_data
-    std::vector<T*> outputs;
+    std::vector<T*> outputs(n);
     auto out_var_names = ctx.OutputNames(framework::GradVarName("X"));
     for (size_t j = 0; j < dx.size(); ++j) {
       if (out_var_names[j] != framework::kEmptyVarName &&
           dx[j]->numel() != 0UL) {
         T* ptr = dx[j]->mutable_data<T>(ctx.GetPlace());
-        outputs.push_back(ptr);
+        outputs[j] = ptr;
       } else {
-        outputs.push_back(nullptr);
+        outputs[j] = nullptr;
       }
     }
     auto dy_data = dy->data<T>();
@@ -195,11 +164,10 @@ class StackGradGPUKernel : public framework::OpKernel<T> {
                  reinterpret_cast<void*>(outputs.data()),
                  outputs.size() * sizeof(T*), dev_ctx.stream());
 
-    int block_dims;
-    int grid_dims;
-    Get1DGpuLaunchConfig(dev_ctx, dy_pre * split_dim * dy_suf, &block_dims,
-                         &grid_dims);
-    unstack_kernel<T><<<grid_dims, block_dims, 0, dev_ctx.stream()>>>(
+    auto config = Get1DGpuLaunchConfig(dev_ctx, dy_pre * split_dim * dy_suf);
+
+    unstack_kernel<T><<<config.block_per_grid, config.thread_per_block, 0,
+                        dev_ctx.stream()>>>(
         dy_data, dy_pre, split_dim, dy_suf, split_dim,
         reinterpret_cast<T**>(tmp_out_data->ptr()));
   }