… paddle/fluid/platform/device_context.h
… paddle/fluid/platform/gpu_info.cc
… paddle/fluid/platform/gpu_info.h