stack's wait cost a lot of cpu time, use cuda kernel to do memory copy
will reduce cpu time.

performance data:
Tensor dim=[2000, 1024, 128]


|tensor number|before opt time cost(ms)|after opt time cost(ms)|
|-- | -- | --|
|2|0.441|0.427
|17|6.7|6.18