@zhaoyuchen2018 2020-01-07
Close this PR because MAC_PR_CI_python35 failed and suspect too many comments may affect CI.
New PR refer to: https://github.com/PaddlePaddle/Paddle/pull/22142